Handle async loading.Initialize plant data as empty array and load the data asynchronously after 2 sec delay
The second parameter has to useEffect as a dependency array, telling the component whether it should execute or
useEffect function when the component renders. If we leave that out, on every rerender of
this component, like from click events, our delay and setSpeakersData will get called.
To tell useEffect to only run its function when the component first renders, pass in an empty array, that is no dependencies.